updated the refactoring
