+++ examples/minibatch_bbh_aynsc/run_bigbench_trace_async.py
@@ -10,7 +10,7 @@ from tqdm import tqdm
 import autogen
 import pickle
 import os
-from opto.trainer.examples.basic_algorithms import MinibatchAlgorithm, evaluate
+from examples.trainers.basic_algorithms import MinibatchAlgorithm, evaluate
 from opto.trainer.guide import Guide
