+++ examples/trainers/UCBsearch.py
@@ -6,7 +6,7 @@ from typing import Union, List, Tuple, Dict, Any, Optional
 from opto import trace
 from opto.trainer.utils import async_run # Assuming print_color is in utils
 from opto.optimizers.utils import print_color
-from opto.trainer.examples.basic_algorithms import MinibatchAlgorithm, evaluate, batchify # evaluate and batchify might be useful
+from examples.trainers.basic_algorithms import MinibatchAlgorithm, evaluate, batchify # evaluate and batchify might be useful
 
 class UCBSearchAlgorithm(MinibatchAlgorithm):
     """
