+++ examples/trainers/__init__.py
@@ -1,0 +1,4 @@
+from examples.trainers.basic_algorithms import Minibatch, MinibatchAlgorithm, BasicSearchAlgorithm
+from examples.trainers.beamsearch_algorithm import BeamsearchAlgorithm, BeamsearchHistoryAlgorithm
+from examples.trainers.UCBsearch import UCBSearchAlgorithm
+from examples.trainers.gepa_algorithms import GEPAAlgorithmBase, GEPAUCBSearch, GEPABeamPareto
