--- opto/features/gepa/__init__.py
@@ -1,9 +1,0 @@
-"""GEPA (Genetic Enhancement via Population Algorithm) implementations.
-
-This module contains experimental GEPA algorithms that extend basic optimization
-with population-based genetic enhancement techniques.
-"""
-
-from .gepa_algorithms import (GEPAAlgorithmBase, GEPAUCBSearch, GEPABeamPareto)
-
-__all__ = ['GEPAAlgorithmBase', 'GEPAUCBSearch', 'GEPABeamPareto']
