+++ opto/trainer/algorithms/__init__.py
@@ -2,3 +2,4 @@ from opto.trainer.algorithms.algorithm import Trainer
 from opto.trainer.algorithms.priority_search import PrioritySearch
 from opto.trainer.algorithms.polca import POLCA
 from opto.trainer.algorithms.streaming_priority_search import StreamingPrioritySearch
+from opto.trainer.algorithms.classical_algorithms import SequentialUpdate, SequentialSearch, BeamSearch
