+++ opto/trainer/algorithms/classical_algorithms.py
