+++ opto/trainer/algorithms/priority_search.py
@@ -6,10 +6,12 @@ from typing import Union, List, Tuple, Dict, Any, Optional, Callable
 from opto import trace
 from opto.trace.nodes import ParameterNode
 from opto.optimizers.optimizer import Optimizer
-from opto.trainer.utils import async_run, safe_mean
-from opto.trainer.examples.basic_algorithms import batchify
-from opto.trainer.supporting_files.search_template import SearchTemplate, Samples, BatchRollout, save_train_config
-from opto.trainer.supporting_files.utils import set_module_parameters, remap_update_dict, create_module_from_update_dict, is_module_copy, deepcopy_module
+from opto.trainer.utils import (
+    async_run, safe_mean, batchify,
+    set_module_parameters, remap_update_dict, create_module_from_update_dict,
+    is_module_copy, deepcopy_module,
+)
+from opto.trainer.search_template import SearchTemplate, Samples, BatchRollout, save_train_config
 from opto.trainer.objectives import (
     ObjectiveConfig, to_score_dict, apply_minimize,
     weighted_scalarize, pareto_rank, aggregate_score_dicts
