+++ opto/trainer/algorithms/streaming_priority_search.py
@@ -1,7 +1,7 @@
 
 import numpy as np
 from typing import List
-from opto.trainer.supporting_files.search_template import Samples, BatchRollout, save_train_config
+from opto.trainer.search_template import Samples, BatchRollout, save_train_config
 from opto.trainer.algorithms.priority_search import PrioritySearch, ModuleCandidate
