+++ opto/trainer/summarizer.py
